requirements.txt
streamlit==1.29.0
pandas==2.2.0
python-docx==1.1.0
reportlab==4.0.7
pillow==10.1.0
